Removes define of CLEF_DEBUG from plugin.
This is really a define that should be in wp-config.php for the specific site. I recommend removing it from the core plugin. That way someone can enable it and not have to worry about modifying core plugin files.

Other benefit - you can keep it enabled across the dev cycle, and updates.